# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GROFO API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspace-layout/{contextKey}:
    parameters:
      - name: contextKey
        in: path
        required: true
        schema:
          type: string
        description: >
          Client-generated key identifying what the layout applies to,
          e.g. "video:12" or "material:5".
    get:
      operationId: getWorkspaceLayout
      tags: [workspace]
      summary: Get the saved split-screen layout for a context
      responses:
        "200":
          description: Layout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceLayout"
        "404":
          description: No saved layout
    put:
      operationId: saveWorkspaceLayout
      tags: [workspace]
      summary: Save/update the split-screen layout for a context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceLayoutInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceLayout"
...
components:
  schemas:
...
    WorkspaceLayoutType:
      type: string
      enum: [video-notes, video-pdf, pdf-notes, ai-pdf, ai-notes]
    WorkspaceLayout:
      type: object
      required: [contextKey, layoutType, splitRatio, noteId, updatedAt]
      properties:
        contextKey:
          type: string
        layoutType:
          $ref: "#/components/schemas/WorkspaceLayoutType"
        splitRatio:
          type: integer
        noteId:
          type: ["integer", "null"]
        updatedAt:
          type: string
          format: date-time
    WorkspaceLayoutInput:
      type: object
      required: [layoutType, splitRatio]
      properties:
        layoutType:
          $ref: "#/components/schemas/WorkspaceLayoutType"
        splitRatio:
          type: integer
          minimum: 0
          maximum: 100
        noteId:
          type: ["integer", "null"]
...